[os] moved os dependencies to test_implementations
